GafferOSL : Replace OSLRenderer with ShadingEngine.
We're phasing out the IECore::Renderer base class, and it really wasn't convenient anyway.

Incompatibilities :

- Replaced OSLRenderer class with ShadingEngine.
